Fix AppBar for both fragments, add two strings for datePicker and img, add date picker calendar
