+++ app/src/main/java/com/peakDevCol/gympartner/ui/home/HomeActivity.kt
@@ -1,17 +1,18 @@
 package com.peakDevCol.gympartner.ui.home
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import androidx.activity.OnBackPressedCallback
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
 import com.peakDevCol.gympartner.R
-import com.peakDevCol.gympartner.core.ex.toast
+import com.peakDevCol.gympartner.core.dialog.BasicDialog
 import com.peakDevCol.gympartner.databinding.ActivityHomeBinding
 import com.peakDevCol.gympartner.domain.ProviderTypeBodyPart
 import dagger.hilt.android.AndroidEntryPoint
@@ -20,8 +21,8 @@ import dagger.hilt.android.AndroidEntryPoint
 class HomeActivity : AppCompatActivity(), OnItemHero {
 
     private lateinit var binding: ActivityHomeBinding
-    private var backPressedOnce = false
-    private val backPressTimeLimit = 2000L // 2 seconds
+    private lateinit var navController: NavController
+    private val homeViewModel: HomeViewModel by viewModels()
 
     companion object {
         fun create(context: Context) = Intent(context, HomeActivity::class.java)
@@ -38,21 +39,40 @@ class HomeActivity : AppCompatActivity(), OnItemHero {
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
-        onBackPressedDispatcher.addCallback(this, onBackPressedCallback)
+        navController =
+            (supportFragmentManager.findFragmentById(R.id.fragment_container) as NavHostFragment).navController
+        initListeners()
     }
 
-    private val onBackPressedCallback = object : OnBackPressedCallback(true) {
-        override fun handleOnBackPressed() {
-            if(backPressedOnce){
-                finish()
-            }
-            backPressedOnce = true
-            toast(getString(R.string.press_back_again_to_exit))
+    private fun initListeners() {
+        binding.topAppBar.setOnMenuItemClickListener { menuItem ->
+            when (menuItem.itemId) {
+                R.id.perfil -> {
+                    navController.navigate(R.id.profileFragment)
+                    true
+                }
 
-            // Reset the backPressedOnce flag after 2 seconds
-            Handler(Looper.getMainLooper()).postDelayed({
-                backPressedOnce = false
-            }, backPressTimeLimit)
+                R.id.logout -> {
+                    showExitDialog()
+                    true
+                }
+
+                else -> false
+            }
+        }
+    }
+
+    @SuppressLint("UseCompatLoadingForDrawables")
+    private fun showExitDialog() {
+        BasicDialog.create(
+            this,
+            resources.getDrawable(R.drawable.dialog_bg),
+            resources.getString(R.string.title_close_session),
+            resources.getString(R.string.supporting_text_close_session),
+            resources.getString(R.string.accept_close_session)
+        ) {
+            homeViewModel.logOut()
+            it.dismiss()
         }
     }
