+++ app/src/main/res/layout/fragment_home.xml
@@ -11,25 +11,6 @@
         android:layout_height="match_parent"
         android:background="@color/background">
 
-        <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/appBar"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:background="@color/primary"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <com.google.android.material.appbar.MaterialToolbar
-                android:id="@+id/topAppBar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:minHeight="?attr/actionBarSize"
-                app:menu="@menu/top_app_bar"
-                app:title="@string/app_name" />
-
-        </com.google.android.material.appbar.AppBarLayout>
-
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/carousel_recycler_view"
             android:layout_width="match_parent"
@@ -42,7 +23,7 @@
             app:layout_constraintBottom_toTopOf="@+id/target"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/appBar" />
+            app:layout_constraintTop_toTopOf="parent" />
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guideline"
@@ -56,7 +37,7 @@
             android:layout_width="200dp"
             android:layout_height="200dp"
             android:layout_margin="8dp"
-            android:contentDescription="no"
+            android:contentDescription="@string/ejercice_img"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/guideline" />
 
@@ -64,7 +45,6 @@
             android:id="@+id/target"
             style="@style/TextInformation"
             android:layout_width="0dp"
-            android:text="lats"
             android:layout_height="wrap_content"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toEndOf="@id/imageView"
